Update Navbar.jsx
Created a reusable NavLink component for mobile menu items to reduce redundancy.
Removed unnecessary md:border class in the mobile menu.
Simplified the logic for toggling the mobile menu, reducing repetitive code. 
I did all the changes that were mentioned in the issue number #21.
Have a look at the changes if anything is require to be changed, please let me know for the same.